pyproject.toml
[tool.poetry]
name = "library_comms"
version = "0.1.0"
description = ""
authors = ["jonathan.kerr <3410350+jonodrew@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.4.2"
freezegun = "^1.5.5"
twilio = "^9.8.3"


[tool.poetry.group.infra.dependencies]
aws-cdk-lib = "^2.222.0"
constructs = "^10.4.2"


[tool.poetry.group.dev.dependencies]
boto3 = "^1.40.65"
aws-cdk-aws-lambda-python-alpha = "^2.222.0a0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

